refactor : API call
